refactor: icons directory
