+++ src/assets/icons/IconBack.tsx
@@ -6,6 +6,7 @@ const IconBack = ({ width, height }: IconProps) => {
       width={width}
       height={height}
       viewBox="0 0 15 14"
+      className="shrink-0"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
