+++ src/assets/icons/IconClose.tsx
@@ -6,6 +6,7 @@ const IconClose = ({ width, height }: IconProps) => {
       width={width}
       height={height}
       viewBox="0 0 17 17"
+      className="shrink-0"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
