+++ src/assets/icons/IconEditPen.tsx
@@ -6,6 +6,7 @@ const IconEditPen = ({ width, height, color }: IconProps) => {
       width={width}
       height={height}
       viewBox="0 0 16 18"
+      className="shrink-0"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
