+++ src/assets/icons/IconFullScreenOn.tsx
@@ -6,6 +6,7 @@ const IconFullScreenOn = ({ width, height }: IconProps) => {
       width={width}
       height={height}
       viewBox="0 0 21 21"
+      className="shrink-0"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
