+++ src/assets/icons/IconNote.tsx
@@ -6,6 +6,7 @@ const IconNote = ({ width, height, color }: IconProps) => {
       width={width}
       height={height}
       viewBox="0 0 22 24"
+      className="shrink-0"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
@@ -23,20 +24,8 @@ const IconNote = ({ width, height, color }: IconProps) => {
         strokeLinecap="round"
         strokeLinejoin="round"
       />
-      <path
-        d="M6 13.0766H12.12"
-        className={color}
-        strokeWidth="1.6"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M6 17.3832H10.08"
-        className={color}
-        strokeWidth="1.6"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
+      <path d="M6 13.0766H12.12" className={color} strokeWidth="1.6" strokeLinecap="round" strokeLinejoin="round" />
+      <path d="M6 17.3832H10.08" className={color} strokeWidth="1.6" strokeLinecap="round" strokeLinejoin="round" />
     </svg>
   )
 }
