+++ src/assets/icons/IconPin.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from '../../../types/iconProps'
+import { IconProps } from '../../types/iconProps'
 
 const IconPin = ({ width, height, color }: IconProps) => {
   return (
