+++ src/assets/icons/IconSettings.tsx
@@ -6,6 +6,7 @@ const IconSettings = ({ width, height }: IconProps) => {
       width={width}
       height={height}
       viewBox="0 0 24 22"
+      className="shrink-0"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
