+++ src/assets/logo.tsx
@@ -4,6 +4,7 @@ const LogoSVG = () => {
       width="222"
       height="46"
       viewBox="0 0 222 46"
+      className="shrink-0"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
